npr-rl/requirements_local.txt
# requirements_local.txt - Simplified dependencies for single-node execution
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy<2.0.0
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
# ray[default]  <-- Removed for local execution
tensordict>=0.8.0,<=0.9.1,!=0.9.0
torchdata
transformers
# vllm==0.8.4
wandb
packaging>=20.0
uvicorn
fastapi
latex2sympy2_extended
math_verify
